calibration/requirements.txt
numpy==1.12.1
opencv-python==3.2.0.7
PyYAML==3.12
